old/requirements.txt
matplotlib
pyqt5
engineering_notation
hurry.filesize
